@@ -20,7 +20,7 @@
     "less": "^4.2.0",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
-    "typescript": "^5.0.2",
+    "typescript": "^5.6.3",
     "vite": "^4.4.5"
   }
 }
